refactor(cli): /auth 슬래시 명령 완전 제거 — /login source 신설, 모든 surface /login 으로 흡수
사용자 명시 결정 (옵션 A): /auth 와 /login 의 cognitive overhead 해소.
/auth 의 잔존 surface (add/remove/set) 모두 /login 으로 옮기고, 명령
자체 + UI/백엔드 + tests 일괄 정리.

핵심 변경:
- /login source <provider> <type> 신규 — 기존 /auth set 의 credential
  source picker. _format_credential_source_label / _persist_credential_source
  / _VALID_* 상수 모두 login.py 로 owner 이전.
- /auth 명령 제거 — routing.py CommandSpec, dispatcher.py cmd_auth
  dispatch + import, core/cli/__init__.py 의 TTY_LOCAL_COMMANDS,
  _state.py 의 COMMAND_MAP + help line, commands/__init__.py 의 export.
- core/cli/commands/auth.py 파일 자체 제거.
- manage_auth LLM tool — backwards-compat adapter 로 유지. 호출 시
  sub_action 을 manage_login 의 subcommand 로 매핑 (legacy prompts 호환).
- core/config/_settings.py + core/wiring/container.py 의 docstring
  /auth → /login source 갱신.
- tests/test_signal_reload.py — gating 검사 의 ("/login", "/key", "/auth")
  → ("/login", "/key").

검증:
- ruff check + format: ✅
- mypy core/cli/commands/ + tool_handlers: ✅
- pytest tests/test_login_command + test_oauth_login + test_manage_login_tool +
  test_signal_reload: 46 pass

호환성:
- /auth → "Unknown command" (slash dispatch). /login 권유 hint
- /auth add / remove / set — 미존재. /login add / remove / source 사용
- manage_auth (LLM tool) — manage_login 으로 자동 forward

후속 (사용자 의도): Anthropic vs OpenAI provider login flow 비교 + 정렬
mismatch 식별. 별도 round-trip 으로 진행.
