+++ core/cli/__init__.py
@@ -139,7 +139,7 @@ def _handle_memory_action(intent: Any, user_text: str, is_offline: bool) -> None
 _LOCAL_COMMANDS = frozenset({"/help"})
 
 # Commands that need TTY interaction locally, then relay the result to serve
-_TTY_LOCAL_COMMANDS = frozenset({"/model", "/auth"})
+_TTY_LOCAL_COMMANDS = frozenset({"/model"})
 
 
 def _thin_interactive_loop(
@@ -264,7 +264,7 @@ def _thin_interactive_loop(
                         break
                     # Notify daemon to reload auth state if this command
                     # may have written to ~/.geode/auth.toml.
-                    if cmd in ("/login", "/key", "/auth"):
+                    if cmd in ("/login", "/key"):
                         import contextlib
 
                         with contextlib.suppress(Exception):
